turtlebotApi/requirements.txt
action-tutorials-py==0.9.4
ament-copyright==0.9.8
ament-cppcheck==0.9.8
ament-cpplint==0.9.8
ament-flake8==0.9.8
ament-index-python==1.1.0
ament-lint==0.9.8
ament-lint-cmake==0.9.8
ament-package==0.9.5
ament-pep257==0.9.8
ament-uncrustify==0.9.8
ament-xmllint==0.9.8
asgiref==3.8.1
backports.zoneinfo==0.2.1
demo-nodes-py==0.9.4
Django==4.2.16
djangorestframework==3.15.2
domain-coordinator==0.9.2
examples-rclpy-executors==0.9.4
examples-rclpy-minimal-action-client==0.9.4
examples-rclpy-minimal-action-server==0.9.4
examples-rclpy-minimal-client==0.9.4
examples-rclpy-minimal-publisher==0.9.4
examples-rclpy-minimal-service==0.9.4
examples-rclpy-minimal-subscriber==0.9.4
launch==0.10.10
launch-ros==0.11.7
launch-testing==0.10.10
launch-testing-ros==0.11.7
launch-xml==0.10.10
launch-yaml==0.10.10
numpy==1.24.4
osrf-pycommon==0.1.11
pika==1.3.2
quality-of-service-demo-py==0.9.4
ros2action==0.9.13
ros2bag==0.3.11
ros2cli==0.9.13
ros2component==0.9.13
ros2controlcli==0.11.0
ros2doctor==0.9.13
ros2interface==0.9.13
ros2launch==0.11.7
ros2lifecycle==0.9.13
ros2multicast==0.9.13
ros2node==0.9.13
ros2param==0.9.13
ros2pkg==0.9.13
ros2run==0.9.13
ros2service==0.9.13
ros2topic==0.9.13
rosidl-runtime-py==0.9.1
rpyutils==0.2.0
rqt-action==1.0.2
rqt-console==1.1.2
rqt-graph==1.1.3
rqt-gui==1.1.2
rqt-gui-py==1.1.2
rqt-msg==1.1.1
rqt-plot==1.1.1
rqt-publisher==1.3.0
rqt-py-console==1.0.2
rqt-reconfigure==1.0.8
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-top==1.0.2
rqt-topic==1.3.0
sqlparse==0.5.1
sros2==0.9.5
teleop-twist-keyboard==2.3.2
topic-monitor==0.9.4
turtlebot3-example==2.1.1
turtlebot3-teleop==2.1.1
typing-extensions==4.12.2
